changing DB datatype from Long to int to save some bytes

+++ src/main/java/redis/clients/jedis/BinaryClient.java
@@ -36,7 +36,7 @@ public class BinaryClient extends Connection {
 
   private String password;
 
-  private long db;
+  private int db;
 
   private boolean isInWatch;
 
@@ -927,7 +927,7 @@ public class BinaryClient extends Connection {
     sendCommand(GETRANGE, key, toByteArray(startOffset), toByteArray(endOffset));
   }
 
-  public Long getDB() {
+  public int getDB() {
     return db;
   }
